PyMongo 3.1 is required
It turns out that PyMongo 3.0.x has a serious issue with threads and
python 2. This has been fixed in PyMongo 3.1rc0.

Also see https://jira.mongodb.org/browse/PYTHON-983

PyMongo 3.1 is not publicy available yet, so we'll just have to wait for
the PyMongo 3.1 release.

If you want to test this branch locally just clone the 3.1rc0 from
github at

https://github.com/mongodb/mongo-python-driver/

and pip install -e from it.